Do not assume MCI_OPEN is always passed with media flags
MCI_OPEN can be used to open a device without any medium. Therefore,
checking the absence of MCI_OPEN_ELEMENT is not a proper method. Check
the flags not supported exactly, instead.

And open a midi file with kmdecOpen() only if MCI_OPEN_ELEMENT is set.
Otherwise, kmdecOpen() fails because a file name is NUL.

This is the observed behvaior when playing a midi by clicking an icon
on WPS.

    modified:   mcdopen.c