Use DosAllocMem() for large memory blocks
kLIBC can manage heaps well, but it does not return unused memory to the
OS until a process terminates.

Or kLIBC allows it manually with _heapmin(), _theapmin() and _uheapmin().
Unfortunately, however, they do not work in DLLs. Need to investigate.

Instead, using DosAllocMem() enables to return unused memory to the OS.

This is needed for very large SF2 files(for example, about 400MB file)
because once it's loaded, the memory allocated for it is never returned to
the OS until system reboot.

    modified:   malloc.c