Fix dead-lock when trying to play
kaiPlay() calls callback to fill buffers initially. BTW, callback is
called in another thread. At this time, hmtxAccessSem is already locked.
Because of this, callback requesting hmtxAccessSem leads to dead-lock.

In case of DART mode, this dead-lock does not occur because callback
is called in the same thread. However, in case of UNIAUD mode and soft
mixer mode this dead-lock occurs because they call callback in other
thread than kaiPlay().

    modified:   mcdopen.c
    modified:   mcdplay.c
    modified:   mcdtemp.h