Fix dead-lock when trying to stop playing a midi file on WPS
In server mode, callback is called in other thread, and when calling
kaiClose() on DART mode, MCI_CLOSE seems to call a mixer event handler
internally. This leads to dead-lock.

kaiStop() is also related to a callback, so suffers from the same
problem.

Check a flag variable once is not enough because if mciDriverEntry()
acquires a mutex and a callback tries to acquire it, then checking a flag
variable only once has no effects. Instead, check it repeatedly.

    modified:   mcdclose.c
    modified:   mcdopen.c
    modified:   mcdplay.c
    modified:   mcdstop.c
    modified:   mcdtemp.h